Allow seek during writes
This allows library users to seek within the written portion of a
file and re-write previously written data. This also includes the
header, which removes the need to specify file size in advance if
the application can tolerate seeking back to update the header at
the end of writing, once the actual file size is known.